+++ sdmx-dl-api/src/main/java/sdmxdl/web/WebSources.java
@@ -8,6 +8,8 @@ import java.util.List;
 @lombok.Builder(toBuilder = true)
 public class WebSources implements HasPersistence {
 
+    public static final WebSources EMPTY = WebSources.builder().build();
+
     @lombok.NonNull
     @lombok.Singular
     List<WebSource> sources;