+++ sdmx-dl-desktop/src/main/java/sdmxdl/desktop/Main.java
@@ -2,9 +2,11 @@ package sdmxdl.desktop;
 
 import com.formdev.flatlaf.FlatLightLaf;
 import ec.util.various.swing.BasicSwingLauncher;
-import internal.sdmxdl.desktop.DesktopWebFactory;
+import sdmxdl.web.SdmxWebManager;
+import sdmxdl.web.WebSource;
 
 import javax.swing.*;
+import java.io.IOException;
 
 @lombok.experimental.UtilityClass
 public class Main {
@@ -20,7 +22,36 @@ public class Main {
 
     static JComponent create() {
         MainComponent mainComponent = new MainComponent();
-        mainComponent.setSdmxManager(DesktopWebFactory.loadManager());
+        mainComponent.setSdmxManager(loadManager());
         return mainComponent;
+    }
+
+    private static SdmxWebManager loadManager() {
+        System.setProperty("enableRngDriver", "true");
+        System.setProperty("enableFileDriver", "true");
+        System.setProperty("enablePxWebDriver", "true");
+        return SdmxWebManager.ofServiceLoader()
+                .toBuilder()
+                .onEvent(Main::printEvent)
+                .onError(Main::printError)
+                .onRegistryEvent(Main::printRegistryEvent)
+                .onRegistryError(Main::printRegistryError)
+                .build();
+    }
+
+    private static void printEvent(WebSource source, String marker, CharSequence message) {
+        System.out.println("[" + source.getId() + "] (" + marker + ") " + message);
+    }
+
+    private static void printError(WebSource source, String marker, CharSequence message, IOException error) {
+        System.err.println("[" + source.getId() + "] (" + marker + ") " + message + ": " + error.getMessage());
+    }
+
+    private static void printRegistryEvent(CharSequence message) {
+        System.out.println("[REG] (" + "" + ") " + message);
+    }
+
+    private static void printRegistryError(CharSequence message, IOException error) {
+        System.err.println("[REG] (" + "" + ") " + message + ": " + error.getMessage());
     }
 }