+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - ![API] Move FileFormat and Persistence to API
+- ![API] Add registry API to load custom web sources
 - ![PROVIDER] Add support of databases in PxWebDriver
 - ![PROVIDER] Add parsing of language in PxWebDriver
 - ![GRPC] Add webservice transcoding in gRPC